1. fixed some bugs

+++ model/commentModel/comment.go
@@ -5,7 +5,7 @@ import "time"
 type Comment struct {
 	Cid       string    `json:"cid" gorm:"cid"`
 	Nid       string    `json:"nid" gorm:"nid"`
-	Uid       int       `json:"uid" gorm:"uid"`
+	Uid       int64     `json:"uid" gorm:"uid"`
 	Content   string    `json:"content" gorm:"content"`
 	ParentId  string    `json:"parent_id" gorm:"parent_id;default:null"`
 	RootId    string    `json:"root_id" gorm:"root_id"`
