+++ model/commentModel/commentDetail.go
@@ -5,7 +5,7 @@ import "time"
 type CommentDetail struct {
 	Cid            string          `json:"cid" gorm:"cid"`
 	Nid            string          `json:"nid" gorm:"nid"`
-	Uid            int             `json:"uid" gorm:"uid"`
+	Uid            int64           `json:"uid" gorm:"uid"`
 	Username       string          `json:"username" gorm:"username"`
 	AvatarUrl      string          `json:"avatarUrl" gorm:"avatar_url"`
 	Content        string          `json:"content" gorm:"content"`
@@ -13,6 +13,6 @@ type CommentDetail struct {
 	ParentUsername string          `json:"parent_name" gorm:"parent_username"`
 	RootId         string          `json:"root_id" gorm:"root_id"`
 	CreatedAt      time.Time       `json:"created_at" gorm:"created_at"`
-	LikesCount     int             `json:"likes_count" gorm:"likes_count"`
+	LikesCount     int64           `json:"likes_count" gorm:"likes_count"`
 	Children       []CommentDetail `json:"children" gorm:"-"`
 }
