+++ model/commentModel/likedComment.go
@@ -1,7 +1,7 @@
 package commentModel
 
 type LikedComment struct {
-	Uid int    `json:"uid" gorm:"uid"`
+	Uid int64  `json:"uid" gorm:"uid"`
 	Cid string `json:"cid" gorm:"cid"`
 }
