+++ model/mqMessageModel/likesNoteComment.go
@@ -9,7 +9,7 @@ import (
 type LikeNoteComment struct {
 	Action    string
 	Cid       string
-	Uid       int
+	Uid       int64
 	Timestamp time.Time
 }
