+++ model/userModel/userCreationInfo.go
@@ -1,12 +1,12 @@
 package userModel
 
 type UserCreationInfo struct {
-	Uid       uint `json:"uid" gorm:"column:uid; default:null"`
-	Follows   uint `json:"follows" gorm:"column:follows; default:null"`
-	Followers uint `json:"followers"  gorm:"column:followers;default:null"`
-	Likes     uint `json:"likes" gorm:"column:likes;default:null"`
-	Collects  uint `json:"collects" gorm:"column:collects;default:null"`
-	NoteCount uint `json:"noteCount" gorm:"column:noteCount;default:null"`
+	Uid       int64 `json:"uid" gorm:"column:uid; default:null"`
+	Follows   int64 `json:"follows" gorm:"column:follows; default:null"`
+	Followers int64 `json:"followers"  gorm:"column:followers;default:null"`
+	Likes     int64 `json:"likes" gorm:"column:likes;default:null"`
+	Collects  int64 `json:"collects" gorm:"column:collects;default:null"`
+	NoteCount int64 `json:"noteCount" gorm:"column:noteCount;default:null"`
 }
 
 func (UserCreationInfo) TableName() string {
